+++ classes/Event/SubmissionsBeforeSendConfirmationNotificationEvent.php
@@ -8,20 +8,24 @@ use Symfony\Contracts\EventDispatcher\Event;
 
 class SubmissionsBeforeSendConfirmationNotificationEvent extends Event
 {
-    /** @var SubmissionModel  */
+    /** @var SubmissionModel */
     private $submission;
 
-    /** @var array  */
+    /** @var array */
     private $submissionCache;
 
-    /** @var FormModel  */
+    /** @var FormModel */
     private $formModel;
 
-    public function __construct(SubmissionModel $submission, array $submissionCache, FormModel $formModel)
+    /** @var array */
+    private $submissionData;
+
+    public function __construct(SubmissionModel $submission, array $submissionCache, FormModel $formModel, array $submissionData)
     {
         $this->submission = $submission;
         $this->submissionCache = $submissionCache;
         $this->formModel = $formModel;
+        $this->submissionData = $submissionData;
     }
 
     /**
@@ -47,4 +51,36 @@ class SubmissionsBeforeSendConfirmationNotificationEvent extends Event
     {
         return $this->formModel;
     }
+
+    /**
+     * Additional value will be added to the submission data for notification
+     *
+     * @return array
+     */
+    public function getSubmissionData(): array
+    {
+        return $this->submissionData;
+    }
+
+    /**
+     * Additional value will be added to the submission data for notification
+     *
+     * @param array $submissionData
+     */
+    public function setSubmissionData(array $submissionData): void
+    {
+        $this->submissionData = $submissionData;
+    }
+
+    /**
+     * Add a single value. If key already exist, it will be overridden.
+     *
+     * Additional value will be added to the submission data for notification
+     */
+    public function addSubmissionData(string $key, string $value): void
+    {
+        $this->submissionData[$key] = $value;
+    }
+
+
 }